Improve data types for almanac parameters

+++ src/core/system_parameters/galileo_almanac_helper.h
@@ -44,7 +44,7 @@ class Galileo_Almanac_Helper
 public:
     // Word type 7: Almanac for SVID1 (1/2), almanac reference time and almanac reference week number
     int32_t IOD_a_7;
-    double WN_a_7;
+    int32_t WN_a_7;
     double t0a_7;
     int32_t SVID1_7;
     double DELTA_A_7;
@@ -59,9 +59,9 @@ public:
     int32_t IOD_a_8;
     double af0_8;
     double af1_8;
-    double E5b_HS_8;
-    double E1B_HS_8;
-    double E5a_HS_8;
+    int32_t E5b_HS_8;
+    int32_t E1B_HS_8;
+    int32_t E5a_HS_8;
     int32_t SVID2_8;
     double DELTA_A_8;
     double e_8;
@@ -72,14 +72,14 @@ public:
 
     // Word type 9: Almanac for SVID2 (2/2) and SVID3 (1/2)
     int32_t IOD_a_9;
-    double WN_a_9;
+    int32_t WN_a_9;
     double t0a_9;
     double M0_9;
     double af0_9;
     double af1_9;
-    double E5b_HS_9;
-    double E1B_HS_9;
-    double E5a_HS_9;
+    int32_t E5b_HS_9;
+    int32_t E1B_HS_9;
+    int32_t E5a_HS_9;
     int32_t SVID3_9;
     double DELTA_A_9;
     double e_9;
@@ -93,9 +93,9 @@ public:
     double M0_10;
     double af0_10;
     double af1_10;
-    double E5b_HS_10;
-    double E1B_HS_10;
-    double E5a_HS_10;
+    int32_t E5b_HS_10;
+    int32_t E1B_HS_10;
+    int32_t E5a_HS_10;
 
     Galileo_Almanac_Helper();  //!< Default constructor
     Galileo_Almanac get_almanac(int i);
